[libromdata] WiiPartitionPrivate::readSector(): Cast sector_num to int64_t to prevent 32-bit overflow in multiplication.
This fixes reading the Wii opening.bnr in games where it's stored after
the 4 GB mark, e.g. "Punch-Out!!" (R7PE01).

+++ src/libromdata/disc/WiiPartition.cpp
@@ -498,7 +498,7 @@ int WiiPartitionPrivate::readSector(uint32_t sector_num)
 
 	// Read the first encrypted sector of the partition.
 	int64_t sector_addr = partition_offset + data_offset;
-	sector_addr += (sector_num * SECTOR_SIZE_ENCRYPTED);
+	sector_addr += ((int64_t)sector_num * SECTOR_SIZE_ENCRYPTED);
 
 	RP_Q(WiiPartition);
 	int ret = discReader->seek(sector_addr);
